# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mural Maz Lince API",
    "description": "API REST del sistema Mural Maz Lince. Digitaliza el tablero de anuncios estudiantil de la Universidad Autónoma de Occidente, Unidad Regional Mazatlán.\n\n**Estado:** Definitivo Sprint 10 — incluye pipeline de moderación (Vision + bad-words + dHash), panel admin y endpoints operativos de shadowban/cola.\n\n**Autenticación:** Cookies HttpOnly con JWT (ADR-04). Incluir `credentials: 'include'` en cada solicitud del cliente.\n\n**Referencia SRS:** v1.4 | **Referencia SAD:** v1.3 | **Referencia DDC:** v1.2",
    "version": "1.0.0-sprint10",
    "contact": {
      "name": "Miriam Alanis — Directora del Proyecto",
      "url": "https://github.com/devHamialco/MuralMazLince"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Entorno local (docker-compose)"
    },
    {
      "url": "https://muralmaz.railway.app",
      "description": "Producción (Railway.app) — disponible al cierre del Sprint 14"
    }
  ],
...
  "paths": {
...
    "/auth/register/student": {
      "post": {
        "tags": ["auth"],
        "summary": "Registro de usuario registrado (ROL-02)",
        "description": "Crea una cuenta de estudiante registrado usando únicamente la matrícula. Valida el formato mediante checksum (8 dígitos, primer dígito = 2). No verifica existencia real en ninguna base de datos externa. **RF-01, RF-07, RN-01, RN-02.**",
        "operationId": "registerStudent",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterStudentRequest" },
              "example": { "matricula": "20240001" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado correctamente. La cookie de sesión se emite automáticamente.",
            "headers": { "Set-Cookie": { "$ref": "#/components/headers/AuthCookie" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" },
                "example": { "role": "visitor_registered", "matricula": "20240001" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/register/entrepreneur": {
      "post": {
        "tags": ["auth"],
        "summary": "Registro de emprendedor (ROL-03)",
        "description": "Crea una cuenta de emprendedor. En una transacción atómica (ADR-09) inserta simultáneamente un registro en `users` y uno en `entrepreneur_profiles`. Si cualquiera falla, ambos se revierten. El `display_name` es obligatorio y se filtra con bad-words. El `whatsapp_number` nunca se expone en ninguna respuesta. **RF-02, RF-39, RF-40, RNF-17, RN-01, RN-02, RN-09.**",
        "operationId": "registerEntrepreneur",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterEntrepreneurRequest" },
              "example": {
                "matricula": "20240001",
                "password": "miPassword123",
                "whatsapp_number": "+5216691234567",
                "display_name": "Mariana G. Tech Lince",
                "privacy_accepted": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Emprendedor registrado. Transacción atómica completada en users + entrepreneur_profiles.",
            "headers": { "Set-Cookie": { "$ref": "#/components/headers/AuthCookie" } },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" },
                "example": { "role": "entrepreneur", "matricula": "20240001", "display_name": "Mariana G. Tech Lince" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
    "/auth/claim-matricula": {
      "post": {
        "tags": ["auth"],
        "summary": "Reclamo de matrícula en uso no autorizado",
        "description": "Crea un ticket de reclamo en la cola del administrador. El sistema no otorga acceso automático a ninguna cuenta hasta la resolución manual del administrador. **RF-35, RN-02, RN-10.**",
        "operationId": "claimMatricula",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ClaimMatriculaRequest" },
              "example": { "matricula": "20240001", "claimant_whatsapp": "+5216691234567" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket de reclamo creado. El administrador contactará al reclamante vía WhatsApp.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "message": "Tu reclamo fue registrado. Un administrador se comunicará contigo vía WhatsApp." }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT en cookie HttpOnly. Emitido por POST /auth/login. No accesible desde JavaScript del cliente (ADR-04). Incluir `credentials: 'include'` en cada solicitud fetch del frontend."
      }
    },
    "headers": {
      "AuthCookie": {
        "description": "Cookie HttpOnly con el JWT de sesión. Gestionada automáticamente por el navegador.",
        "schema": { "type": "string" }
      }
    },
...
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Descripción del error. Nunca expone datos sensibles." }
        }
      },
      "AuthResponse": {
        "type": "object",
        "description": "Respuesta de autenticación. No incluye whatsapp_number ni password_hash en ningún caso.",
        "required": ["role", "matricula"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["visitor_registered", "entrepreneur", "admin"]
          },
          "matricula": { "type": "string", "example": "20240001" },
          "display_name": {
            "type": "string",
            "nullable": true,
            "description": "Presente solo para ROL-03. Proveniente de entrepreneur_profiles.display_name.",
            "example": "Mariana G. Tech Lince"
          }
        }
      },
      "RegisterStudentRequest": {
        "type": "object",
        "required": ["matricula"],
        "properties": {
          "matricula": {
            "type": "string",
            "pattern": "^2[0-9]{7}$",
            "description": "8 dígitos, primer dígito = 2. Checksum local (RN-01).",
            "example": "20240001"
          }
        }
      },
      "RegisterEntrepreneurRequest": {
        "type": "object",
        "required": ["matricula", "password", "whatsapp_number", "display_name", "privacy_accepted"],
        "properties": {
          "matricula": {
            "type": "string",
            "pattern": "^2[0-9]{7}$",
            "example": "20240001"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Mínimo 8 caracteres. Se almacena hasheada con bcrypt (factor 10). RNF-06.",
            "example": "miPassword123"
          },
          "whatsapp_number": {
            "type": "string",
            "pattern": "^\\+52[0-9]{10}$",
            "description": "Formato +52 + 10 dígitos. Nunca se expone en respuestas de la API. RN-09.",
            "example": "+5216691234567"
          },
          "display_name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80,
            "description": "Nombre visible en el feed. Obligatorio. Sujeto a filtro bad-words. RF-39.",
            "example": "Mariana G. Tech Lince"
          },
          "privacy_accepted": {
            "type": "boolean",
            "enum": [true],
            "description": "Aceptación del Aviso de Privacidad (LFPDPPP). Debe ser true para completar el registro. RNF-17."
          }
        }
      },
...
      "ClaimMatriculaRequest": {
        "type": "object",
        "required": ["matricula", "claimant_whatsapp"],
        "properties": {
          "matricula": {
            "type": "string",
            "pattern": "^2[0-9]{7}$",
            "example": "20240001"
          },
          "claimant_whatsapp": {
            "type": "string",
            "pattern": "^\\+52[0-9]{10}$",
            "description": "WhatsApp del reclamante para que el administrador contacte. RF-35.",
            "example": "+5216699876543"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "400 Bad Request — Checksum inválido, campos faltantes o formato de WhatsApp incorrecto.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "El formato de la matrícula es inválido. Debe tener 8 dígitos y comenzar con 2." }
          }
        }
      },
...
      "Conflict": {
        "description": "409 Conflict — Matrícula ya registrada, reporte duplicado o límite de proyectos/anuncios alcanzado.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Esta matrícula ya está registrada en el sistema." }
          }
        }
      },
...
      "InternalServerError": {
        "description": "500 Internal Server Error — Error inesperado en el servidor. El detalle se registra en el log; el cliente recibe un mensaje genérico.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Ocurrió un error interno. Por favor intenta más tarde." }
          }
        }
      }
    }
  }
}